app/pipeline/requirements.txt
pandas==1.5.3
aiohttpaiohttp==3.8.4
playwright==1.34.0
beautifulsoup4==4.12.2
certifi==2023.5.7
openpyxl==3.1.2
sentence-transformers==2.2.2
google-api-python-client==2.84.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.10.0
python-dotenv==1.0.0
pyarrow==11.0.0
subprocess==0.2.0
pyment==0.2.0
loguru==0.6.0